+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # UNRELEASED
 
 **New Features:**
-- add new `touchEventOptions` prop that can set the options for the the touch event listeners
-  - this helps provides users full control of if/when they want to set [passive](https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/addEventListener#options)
+- add new `touchEventOptions` prop that can set the options for the touch event listeners
+  - this provides users full control of if/when they want to set [passive](https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/addEventListener#options)
 - add new `onTouchStartOrOnMouseDown` prop that is called for `touchstart` and `mousedown`. Before a swipe even starts.
   - combined with `touchEventOptions` allows users the ability to now call `preventDefault` on `touchstart`
-- add new `onTouchEndOrOnMouseUp` prop that is always called for `touchend` and `mouseup`.
+- add new `onTouchEndOrOnMouseUp` prop that is called for `touchend` and `mouseup`.
 
 **Breaking Changes:**
 * we have dropped support for `es5` transpiled output